реализовал сохранение контактов в localStorage

+++ client/src/pages/Contacts/index.tsx
@@ -1,29 +1,39 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect, useRef, useState } from "react";
 import ContactInfo from "./ContactInfo";
 import ContactList from "./ContactList";
 import styles from "./Contacts.module.scss";
 import { motion } from "framer-motion";
 import { useDispatch, useSelector } from "react-redux";
-import { setContacts } from "../../redux/contactsSlice";
-import data from "../../assets/data.json";
 import { Link } from "react-router-dom";
+import { cleanItems } from "../../redux/contactsSlice";
 
 const Contacts = () => {
   const dispatch = useDispatch();
   const [editingMode, setEditingMode] = useState(false);
   const [render, setRender] = useState(false);
+  const isMounted = useRef(false);
+
+  // @ts-ignore
+  const contacts = useSelector((state) => state.contacts.items);
 
   useEffect(() => {
-    dispatch(setContacts(data.users));
-  }, []);
+    const json = JSON.stringify(contacts);
+    localStorage.setItem("contacts", json);
+  }, [contacts]);
 
   const getEditingMode = (mode: boolean) => {
     setEditingMode(mode);
   };
 
   const onClickLogOut = () => {
-    localStorage.clear();
-    setRender(!render);
+    const result = window.confirm(
+      "All contacts will be lost. Are you sure you want to log out?"
+    );
+    if (result) {
+      localStorage.clear();
+      dispatch(cleanItems());
+      setRender(!render);
+    }
   };
 
   return (
